Added Registration Page
