+++ src/app/User_RegxLogin/CreateAccount/CreateAccount.module.css
@@ -1,0 +1,203 @@
+.joinUs {
+  	position: absolute;
+  	top: 0px;
+  	left: 0px;
+  	font-weight: 900;
+}
+.logo {
+  	position: absolute;
+  	top: 309.47px;
+  	left: 256px;
+  	width: 198px;
+  	height: 63px;
+  	text-align: left;
+  	font-size: 50px;
+}
+.aboutUsChild {
+  	position: absolute;
+  	top: 89px;
+  	left: 114px;
+  	border-radius: 40px;
+  	background-color: #222121;
+  	width: 200px;
+  	height: 72px;
+}
+.aboutUs1 {
+  	position: absolute;
+  	top: 111px;
+  	left: 155px;
+}
+.forMoreInformation {
+  	position: absolute;
+  	top: 0px;
+  	left: 0px;
+  	font-size: 20px;
+  	display: inline-block;
+  	width: 426px;
+  	height: 65px;
+}
+.aboutUs {
+  	position: absolute;
+  	top: 512.47px;
+  	left: 142px;
+  	width: 426px;
+  	height: 161px;
+  	font-size: 26px;
+}
+.formIcon {
+  	width: 100%;
+	position: relative;
+	top: 145px;
+  	left: 260px;
+	max-width: 100%;
+	overflow: hidden;
+	height: 734.9px;
+	object-fit: contain;
+}
+.register {
+  	position: absolute;
+  	top: 199.47px;
+  	left: 827px;
+  	font-size: 40px;
+  	display: inline-block;
+  	color: #f3e2e2;
+  	width: 343px;
+  	height: 82px;
+}
+.formChild {
+  	position: absolute;
+  	top: 118px;
+  	left: 0px;
+  	box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  	border-radius: 20px;
+  	background-color: #fff;
+  	border: 2px solid rgba(0, 0, 0, 0.25);
+  	box-sizing: border-box;
+  	width: 450px;
+  	height: 55px;
+}
+.formItem {
+  	position: absolute;
+  	top: 0px;
+  	left: 0px;
+  	box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  	border-radius: 20px;
+  	background-color: #fff;
+  	border: 2px solid rgba(0, 0, 0, 0.25);
+  	box-sizing: border-box;
+  	width: 450px;
+  	height: 56px;
+}
+.formInner {
+  	position: absolute;
+  	top: 236px;
+  	left: 0px;
+  	box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  	border-radius: 20px;
+  	background-color: #fff;
+  	border: 2px solid rgba(0, 0, 0, 0.25);
+  	box-sizing: border-box;
+  	width: 450px;
+  	height: 56px;
+}
+.universityEmail {
+  	position: absolute;
+  	top: 13px;
+  	left: 24px;
+}
+.password {
+  	position: absolute;
+  	top: 250px;
+  	left: 40px;
+}
+.courseYear {
+  	position: absolute;
+  	top: 132px;
+  	left: 24px;
+}
+.form {
+  	position: absolute;
+  	top: 292.47px;
+  	left: 759px;
+  	width: 450px;
+  	height: 292px;
+  	color: #c6c3c3;
+}
+.registerChild {
+  	position: absolute;
+  	top: 0px;
+  	left: 0px;
+  	border-radius: 20px;
+  	background-color: #191919;
+  	width: 226px;
+  	height: 56px;
+}
+.register2 {
+  	position: absolute;
+  	top: 11px;
+  	left: 63px;
+}
+.register1 {
+  	position: absolute;
+  	top: 667.47px;
+  	left: 758px;
+  	display: flex;
+  	align-items: center;
+  	justify-content: flex-start;
+  	padding-left: 24px;
+  	width: 184px;
+  	height: 23px;
+  	color: #fff;
+  	font-size: 20px;
+  	font-weight: bold;
+  	cursor: pointer;
+  	border: none;
+  	background: none;
+  	border-radius: 0;
+  	transition: color 0.2s;
+  	text-align: left;
+}
+.register1:hover {
+  	color: #fac36b;
+}
+.logInHere {
+  	position: absolute;
+  	top: 603.47px;
+  	left: 779px;
+  	font-size: 20px;
+  	display: inline-block;
+  	width: 223px;
+  	height: 32px;
+	text-decoration: none;
+  	color: #fff;
+  	font-weight: bold;
+  	cursor: pointer;
+  	border: none;
+  	background: none;
+  	border-radius: 0;
+  	transition: color 0.2s;
+  	text-align: left;
+}
+.logInHere:hover{
+  	color: #fac36b;
+}
+.logoOfMsad1 {
+  	position: absolute;
+  	top: 145.47px;
+  	left: 282px;
+  	width: 147px;
+  	height: 147px;
+  	object-fit: cover;
+}
+.registration {
+  	width: 100%;
+  	position: relative;
+  	background: linear-gradient(180deg, #bb8b1b, #b01818);
+  	height: 1024px;
+  	overflow: hidden;
+  	text-align: center;
+  	font-size: 24px;
+  	color: #fff;
+  	font-family: Merriweather;
+}
+
